fix(sheet): theme the resource bar and move its props type out
The bar carried its border, fill, text and shadow colours as literals,
carried over from the mana bar it generalizes. They now come from the
theme, with light and dark values, so the bar reads correctly in both.

Props types move to types/components/ResourceBar.d.ts, and the maximum
callback is optional rather than being passed an empty function by the
one caller that cannot edit a maximum.
